requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.2.2               # via kombu
asn1crypto==0.24.0        # via cryptography
attrs==17.4.0             # via pytest
babel==2.5.3
billiard==3.5.0.3         # via celery
bleach==2.1.3
boto3==1.5.31
botocore==1.8.45          # via boto3, s3transfer
braintree==3.39.1         # via django-payments
cairocffi==0.8.0          # via cairosvg, weasyprint
cairosvg==2.1.3
celery[redis]==4.2.0
certifi==2018.1.18        # via requests
cffi==1.11.4              # via cairocffi, cryptography, weasyprint
chardet==3.0.4            # via requests
cryptography==2.2.2       # via django-payments
cssselect2==0.2.1         # via cairosvg, weasyprint
defusedxml==0.5.0         # via cairosvg, python3-openid, social-auth-core
dj-database-url==0.4.2
dj-email-url==0.0.10
django-babel==0.6.2
django-bootstrap4==0.0.5
django-cache-url==2.0.0
django-celery-results==1.0.1
django-countries==5.2
django-debug-toolbar-request-history==0.0.7
django-debug-toolbar==1.9.1
django-elasticsearch-dsl==0.4.4
django-environ==0.4.4     # via django-graphql-jwt
django-filter==1.1
django-fsm==2.6.0
django-graphql-jwt==0.1.9
django-impersonate==1.3
django-js-asset==1.0.0    # via django-mptt
django-mptt==0.9.0
django-payments==0.12.3
django-phonenumber-field==2.0.0
django-prices-openexchangerates==1.0.0
django-prices-vatlayer==1.0.1
django-prices==1.0.1
django-recaptcha==1.4.0
django-redis==4.8.0
django-render-block==0.5  # via django-templated-email
django-storages==1.6.5
django-templated-email==2.2.0
django-versatileimagefield==1.9
django-webpack-loader==0.5.0
django==2.0.3
docutils==0.14            # via botocore
elasticsearch-dsl==5.4.0
elasticsearch==5.5.2
faker==0.8.11
freezegun==0.3.10
google-i18n-address==2.3.0
google-measurement-protocol==1.0.0
graphene-django==2.0.0
graphene==2.0.1           # via django-graphql-jwt, graphene-django
graphql-core==2.0         # via graphene, graphql-relay
graphql-relay==0.4.5      # via graphene
html5lib==1.0.1           # via bleach, weasyprint
idna==2.6                 # via cryptography, requests
iso8601==0.1.12           # via graphene-django
jmespath==0.9.3           # via boto3, botocore
jsonfield==2.0.2
kombu==4.2.1              # via celery
markdown==2.6.11
maxminddb-geolite2==2017.803
maxminddb==1.3.0
oauthlib==2.0.6           # via requests-oauthlib, social-auth-core
pdfrw==0.4                # via weasyprint
phonenumberslite==8.9.1
pillow==5.0.0             # via cairosvg, django-versatileimagefield
pluggy==0.6.0             # via pytest
prices==1.0.0
promise==2.1              # via graphene, graphene-django, graphql-core, graphql-relay
psycopg2-binary==2.7.5
purl==1.3.1
py==1.5.2                 # via pytest
pycparser==2.18           # via cffi
pyjwt==1.5.3              # via django-graphql-jwt, django-payments, social-auth-core
pyphen==0.9.4             # via weasyprint
pytest-django==3.1.2
pytest-mock==1.7.0
pytest-vcr==0.3.0
pytest==3.4.0
python-dateutil==2.6.1    # via botocore, elasticsearch-dsl, faker, freezegun
python3-openid==3.1.0     # via social-auth-core
pytz==2018.3              # via babel, celery, django
pyyaml==3.13              # via vcrpy
raven==6.9.0
redis==2.10.6             # via celery, django-redis
requests-oauthlib==0.8.0  # via social-auth-core
requests==2.18.4
rx==1.6.0                 # via graphql-core
s3transfer==0.1.12        # via boto3
singledispatch==3.4.0.3   # via graphene-django
six==1.11.0               # via bleach, cryptography, django-environ, django-templated-email, elasticsearch-dsl, faker, freezegun, graphene, graphene-django, graphql-core, graphql-relay, html5lib, promise, purl, pytest, python-dateutil, singledispatch, social-auth-app-django, social-auth-core, vcrpy
social-auth-app-django==2.1.0
social-auth-core==1.6.0   # via social-auth-app-django
sqlparse==0.2.4           # via django-debug-toolbar
stripe==1.77.1            # via django-payments
suds-jurko==0.6           # via django-payments
text-unidecode==1.1
tinycss2==0.6.1           # via cairosvg, cssselect2, weasyprint
typing==3.6.2             # via google-measurement-protocol, prices, promise
urllib3==1.22             # via elasticsearch, requests
uwsgi==2.0.17 ; platform_system != "Windows"
vcrpy==1.11.1
vine==1.1.4               # via amqp
weasyprint==0.42.2
webencodings==0.5.1       # via html5lib, tinycss2
wrapt==1.10.11            # via vcrpy
xmltodict==0.11.0         # via django-payments
